new try

+++ exercises.c
@@ -118,7 +118,7 @@ int parentesisBalanceados(char *cadena) {
    int i = 0;
    while(cadena[i] != '\0'){
       if (cadena[i] == '(' || cadena[i] == '{' || cadena[i] == '['){
-         push(pila,(void*)(int)cadena[i]);
+         push(pila,(void*)(long)cadena[i]);
       }
       else if(cadena[i] == ')' || cadena[i] == '}' || cadena[i] == ']'){
          if(top(pila) == NULL){
